Fix NON_EQUIVALENT build.

+++ src/waves.h
@@ -119,6 +119,6 @@ f32 func_800BB2F4(s32, f32, f32, f32*);
 void func_800B8C04(s32, s32, s32, s32, s32);
 void func_800BF524(Object *);
 void *func_800BF634(Object *arg0, f32 arg1, f32 arg2, f32 arg3, s32 arg4, f32 arg5, f32 arg6, f32 arg7, s32 arg8);
-f32 func_800BEFC4(u16, u16, u16);
+f32 func_800BEFC4(s32, s32, s32);
 
 #endif
